refactor: replace Html2rss::RssBuilder with modern AutoSource::RssBuilder
RssBuilder functionality has been moved from AutoSource::RssBuilder to Html2rss::RssBuilder. The separate file for AutoSource::RssBuilder was deleted, and the references were updated across the codebase, resulting in a more unified structure. This change simplifies the module hierarchy and improves maintainability by reducing redundancy.
