+++ spec/lib/html2rss/rss_builder_spec.rb
@@ -1,5 +1,5 @@
 # frozen_string_literal: true
 
 RSpec.describe Html2rss::RssBuilder do
-  it { expect(described_class).not_to respond_to(:new) }
+  it { expect(described_class).to be_a Class }
 end
